Fixed issue when secrets.json didn't exists. Set correct working directory for local automation host.

+++ c_sharp_for_home_assistant/src/hhnl.HomeAssistantNet/hhnl.HomeAssistantNet.CSharpForHomeAssistant/Services/SecretsService.cs
@@ -60,7 +60,8 @@ namespace hhnl.HomeAssistantNet.CSharpForHomeAssistant.Services
 
         public void DeploySecretsFile()
         {
-            File.Copy(_secretsConfigPath, _secretsDeployPath, true);
+            if(File.Exists(_secretsConfigPath))
+                File.Copy(_secretsConfigPath, _secretsDeployPath, true);
         }
 
         private class SecretsFile
